+++ components/restreamer/client/cypress/e2e/add_output_single.cy.js
@@ -1,4 +1,4 @@
-describe('ADD SINGLE OUTPUT', () => {
+describe('ADD SINGLE OUTPUT', { testIsolation: false }, () => {
   before(() => {
     cy.visit('/');
     cy.deleteAllInputs();