+++ components/restreamer/src/cli.rs
@@ -2,7 +2,6 @@
 
 use std::{fmt, net::IpAddr, path::PathBuf, str::FromStr as _};
 
-use anyhow::anyhow;
 use ephyr_log::tracing;
 use structopt::StructOpt;
 
@@ -129,11 +128,19 @@ pub struct Opts {
     #[structopt(
         short,
         long,
-        parse(try_from_str = Self::parse_log_level),
-        help = "Logs verbosity level: \
-                OFF | ERRO | WARN | INFO | DEBG | TRCE"
+        parse(try_from_str = tracing::Level::from_str),
+        help = "Logs verbosity level: INFO | DEBUG | TRACE"
     )]
     pub verbose: Option<tracing::Level>,
+
+    /// Logs format for displaying.
+    #[structopt(
+        short,
+        long,
+        env = "EPHYR_RESTREAMER_LOG_FORMAT",
+        help = "Logs format: JSON | COMPACT"
+    )]
+    pub log_format: Option<ephyr_log::LogFormat>,
 
     /// Path for local video files.
     #[structopt(
@@ -146,33 +153,41 @@ pub struct Opts {
     )]
     pub file_root: PathBuf,
 
-    /// Port of Jaeger to send traces.
+    /// IP address of [OpenTelemetry] collector server to send logs to.
+    ///
+    /// [OpenTelemetry]: https://OpenTelemetry.io
     #[structopt(
         long,
-        env = "EPHYR_RESTREAMER_JAEGER_AGENT_IP",
-        help = "IP of Jaeger to send traces",
-        long_help = "Uses for aggregation of traces for Jaeger"
+        env = "EPHYR_RESTREAMER_OTLP_COLLECTOR_IP",
+        help = "IP of OTLP collector to send traces",
+        long_help = "Uses for aggregation of traces for OTLP collector"
     )]
-    pub jaeger_agent_ip: Option<IpAddr>,
+    pub otlp_collector_ip: Option<IpAddr>,
 
-    /// Port of Jaeger to send traces.
+    /// Port of [OpenTelemetry] collector server to send logs to.
+    ///
+    /// In our case as we send data with gRPC so port is typically `4317`.
+    ///
+    /// [OpenTelemetry]: https://OpenTelemetry.io
     #[structopt(
         long,
-        env = "EPHYR_RESTREAMER_JAEGER_AGENT_PORT",
-        help = "Port of Jaeger to send traces",
-        long_help = "Uses for aggregation of traces for Jaeger"
+        env = "EPHYR_RESTREAMER_OTLP_COLLECTOR_PORT",
+        help = "Port of OTLP collector to send traces",
+        long_help = "Uses for aggregation of traces for OTLP collector"
     )]
-    pub jaeger_agent_port: Option<u16>,
+    pub otlp_collector_port: Option<u16>,
 
-    /// Service name for Jaeger to send traces.
+    /// Service name to collect traces to [OpenTelemetry] collector.
+    ///
+    /// [OpenTelemetry]: https://OpenTelemetry.io
     #[structopt(
         long,
-        env = "EPHYR_RESTREAMER_JAEGER_SERVICE_NAME",
+        env = "EPHYR_RESTREAMER_SERVICE_NAME",
         default_value = "ephyr-restreamer",
-        help = "Service name for Jaeger to send traces",
-        long_help = "Uses for aggregation of traces for Jaeger"
+        help = "Service name to collect traces to OTLP collector",
+        long_help = "Uses for aggregation of traces for OTLP collector"
     )]
-    pub jaeger_service_name: String,
+    pub service_name: String,
 }
 
 impl Opts {
@@ -183,29 +198,6 @@ impl Opts {
     #[must_use]
     pub fn from_args() -> Self {
         <Self as StructOpt>::from_args()
-    }
-
-    /// Parses [`tracing::Level`] from the given string.
-    ///
-    /// This function is required, because [`tracing::Level`]'s [`FromStr`]
-    /// implementation returns `()`, which is not [`Display`] as [`StructOpt`]
-    /// requires.
-    ///
-    /// # Errors
-    ///
-    /// If [`tracing::Level`] failed to parse from the string.
-    ///
-    /// [`Display`]: std::fmt::Display
-    /// [`FromStr`]: std::str::FromStr
-    pub fn parse_log_level(lvl: &str) -> Result<tracing::Level, anyhow::Error> {
-        #[allow(clippy::map_err_ignore)]
-        tracing::Level::from_str(lvl).map_err(|_| {
-            anyhow!(
-                "'{}' is invalid verbosity level, allowed levels are: \
-                 OFF | CRIT | ERRO | WARN | INFO | DEBG | TRCE",
-                lvl,
-            )
-        })
     }
 }