+++ components/restreamer/src/lib.rs
@@ -34,6 +34,7 @@ pub mod console_logger;
 pub mod dvr;
 pub mod ffmpeg;
 pub mod file_manager;
+mod proc;
 pub mod serde;
 pub mod server;
 pub mod spec;